[refactor] move login related settings to own subfolder
