+++ src/login/PostLoginActions.ts
@@ -250,7 +250,7 @@ export class PostLoginActions implements IPostLoginAction {
 
 	private enforcePasswordChange(): void {
 		if (logins.getUserController().user.requirePasswordUpdate) {
-			import("../settings/ChangePasswordDialogs.js").then(({showChangeOwnPasswordDialog}) => {
+			import("../settings/login/ChangePasswordDialogs.js").then(({showChangeOwnPasswordDialog}) => {
 				return showChangeOwnPasswordDialog(false)
 			})
 		}
