+++ src/settings/SettingsView.ts
@@ -7,7 +7,7 @@ import {SettingsFolder} from "./SettingsFolder"
 import {lang} from "../misc/LanguageViewModel"
 import type {CurrentView} from "../gui/Header.js"
 import {header} from "../gui/Header.js"
-import {LoginSettingsViewer} from "./LoginSettingsViewer"
+import {LoginSettingsViewer} from "./login/LoginSettingsViewer.js"
 import {GlobalSettingsViewer} from "./GlobalSettingsViewer"
 import {DesktopSettingsViewer} from "./DesktopSettingsViewer"
 import {MailSettingsViewer} from "./MailSettingsViewer"
