+++ src/settings/login/LoginSettingsViewer.ts
@@ -1,37 +1,37 @@
 import m, {Children} from "mithril"
 import stream from "mithril/stream"
-import type {TextFieldAttrs} from "../gui/base/TextField.js"
-import {TextField} from "../gui/base/TextField.js"
-import {InfoLink, lang} from "../misc/LanguageViewModel"
-import {logins} from "../api/main/LoginController"
-import {Icons} from "../gui/base/icons/Icons"
-import {CustomerPropertiesTypeRef, Session, SessionTypeRef} from "../api/entities/sys/TypeRefs.js"
+import type {TextFieldAttrs} from "../../gui/base/TextField.js"
+import {TextField} from "../../gui/base/TextField.js"
+import {InfoLink, lang} from "../../misc/LanguageViewModel.js"
+import {logins} from "../../api/main/LoginController.js"
+import {Icons} from "../../gui/base/icons/Icons.js"
+import {CustomerPropertiesTypeRef, Session, SessionTypeRef} from "../../api/entities/sys/TypeRefs.js"
 import {assertNotNull, LazyLoaded, neverNull, ofClass} from "@tutao/tutanota-utils"
-import {formatDateTimeFromYesterdayOn} from "../misc/Formatter"
-import {SessionState} from "../api/common/TutanotaConstants"
-import {EditSecondFactorsForm} from "./EditSecondFactorsForm"
-import type {EntityUpdateData} from "../api/main/EventController"
-import {isUpdateForTypeRef} from "../api/main/EventController"
-import {NotFoundError} from "../api/common/error/RestError"
-import * as RecoverCodeDialog from "./RecoverCodeDialog"
-import {attachDropdown} from "../gui/base/Dropdown.js"
-import {ExpanderButton, ExpanderPanel} from "../gui/base/Expander"
-import {ColumnWidth, Table} from "../gui/base/Table.js"
-import {ifAllowedTutanotaLinks} from "../gui/base/GuiUtils"
-import type {UpdatableSettingsViewer} from "./SettingsView"
-import {CredentialEncryptionMode} from "../misc/credentials/CredentialEncryptionMode"
-import type {CredentialsProvider} from "../misc/credentials/CredentialsProvider.js"
-import {hasKeychainAuthenticationOptions} from "../misc/credentials/CredentialsProviderFactory"
-import {showCredentialsEncryptionModeDialog} from "../gui/dialogs/SelectCredentialsEncryptionModeDialog"
-import {assertMainOrNode} from "../api/common/Env"
-import {locator} from "../api/main/MainLocator"
-import {elementIdPart, getElementId} from "../api/common/utils/EntityUtils"
+import {formatDateTimeFromYesterdayOn} from "../../misc/Formatter.js"
+import {SessionState} from "../../api/common/TutanotaConstants.js"
+import {EditSecondFactorsForm} from "./secondfactor/EditSecondFactorsForm.js"
+import type {EntityUpdateData} from "../../api/main/EventController.js"
+import {isUpdateForTypeRef} from "../../api/main/EventController.js"
+import {NotFoundError} from "../../api/common/error/RestError.js"
+import * as RecoverCodeDialog from "./RecoverCodeDialog.js"
+import {attachDropdown} from "../../gui/base/Dropdown.js"
+import {ExpanderButton, ExpanderPanel} from "../../gui/base/Expander.js"
+import {ColumnWidth, Table} from "../../gui/base/Table.js"
+import {ifAllowedTutanotaLinks} from "../../gui/base/GuiUtils.js"
+import type {UpdatableSettingsViewer} from "../SettingsView.js"
+import {CredentialEncryptionMode} from "../../misc/credentials/CredentialEncryptionMode.js"
+import type {CredentialsProvider} from "../../misc/credentials/CredentialsProvider.js"
+import {hasKeychainAuthenticationOptions} from "../../misc/credentials/CredentialsProviderFactory.js"
+import {showCredentialsEncryptionModeDialog} from "../../gui/dialogs/SelectCredentialsEncryptionModeDialog.js"
+import {assertMainOrNode} from "../../api/common/Env.js"
+import {locator} from "../../api/main/MainLocator.js"
+import {elementIdPart, getElementId} from "../../api/common/utils/EntityUtils.js"
 import {showChangeOwnPasswordDialog} from "./ChangePasswordDialogs.js";
-import {IconButton, IconButtonAttrs} from "../gui/base/IconButton.js"
-import {ButtonSize} from "../gui/base/ButtonSize.js"
-import {DropDownSelector, DropDownSelectorAttrs} from "../gui/base/DropDownSelector.js"
-import {UsageTestModel} from "../misc/UsageTestModel.js"
-import {UserSettingsGroupRootTypeRef} from "../api/entities/tutanota/TypeRefs.js"
+import {IconButton, IconButtonAttrs} from "../../gui/base/IconButton.js"
+import {ButtonSize} from "../../gui/base/ButtonSize.js"
+import {DropDownSelector, DropDownSelectorAttrs} from "../../gui/base/DropDownSelector.js"
+import {UsageTestModel} from "../../misc/UsageTestModel.js"
+import {UserSettingsGroupRootTypeRef} from "../../api/entities/tutanota/TypeRefs.js"
 
 assertMainOrNode()
