+++ src/settings/login/RecoverCodeDialog.ts
@@ -1,14 +1,14 @@
-import {InfoLink, lang} from "../misc/LanguageViewModel"
-import {Dialog, DialogType} from "../gui/base/Dialog"
+import {InfoLink, lang} from "../../misc/LanguageViewModel.js"
+import {Dialog, DialogType} from "../../gui/base/Dialog.js"
 import type {Hex} from "@tutao/tutanota-utils"
 import {neverNull, noOp, ofClass} from "@tutao/tutanota-utils"
 import m, {Children, Vnode} from "mithril"
-import {assertMainOrNode, isApp} from "../api/common/Env"
-import {copyToClipboard} from "../misc/ClipboardUtils"
-import {Button} from "../gui/base/Button.js"
-import {AccessBlockedError, NotAuthenticatedError} from "../api/common/error/RestError"
-import {locator} from "../api/main/MainLocator"
-import {Icons} from "../gui/base/icons/Icons.js"
+import {assertMainOrNode, isApp} from "../../api/common/Env.js"
+import {copyToClipboard} from "../../misc/ClipboardUtils.js"
+import {Button} from "../../gui/base/Button.js"
+import {AccessBlockedError, NotAuthenticatedError} from "../../api/common/error/RestError.js"
+import {locator} from "../../api/main/MainLocator.js"
+import {Icons} from "../../gui/base/icons/Icons.js"
 
 type Action = "get" | "create"
 assertMainOrNode()
