+++ src/settings/login/secondfactor/EditSecondFactorDialog.ts
@@ -1,34 +1,34 @@
-import {showProgressDialog} from "../gui/dialogs/ProgressDialog.js"
-import {GroupType, SecondFactorType} from "../api/common/TutanotaConstants.js"
-import type {DropDownSelectorAttrs, SelectorItem} from "../gui/base/DropDownSelector.js"
-import {DropDownSelector} from "../gui/base/DropDownSelector.js"
-import type {TranslationKey} from "../misc/LanguageViewModel.js"
-import {lang} from "../misc/LanguageViewModel.js"
-import type {TextFieldAttrs} from "../gui/base/TextField.js"
-import {TextField} from "../gui/base/TextField.js"
-import {isApp, isTutanotaDomain} from "../api/common/Env.js"
+import {showProgressDialog} from "../../../gui/dialogs/ProgressDialog.js"
+import {GroupType, SecondFactorType} from "../../../api/common/TutanotaConstants.js"
+import type {DropDownSelectorAttrs, SelectorItem} from "../../../gui/base/DropDownSelector.js"
+import {DropDownSelector} from "../../../gui/base/DropDownSelector.js"
+import type {TranslationKey} from "../../../misc/LanguageViewModel.js"
+import {lang} from "../../../misc/LanguageViewModel.js"
+import type {TextFieldAttrs} from "../../../gui/base/TextField.js"
+import {TextField} from "../../../gui/base/TextField.js"
+import {isApp, isTutanotaDomain} from "../../../api/common/Env.js"
 import m, {Children} from "mithril"
-import {Button, ButtonType} from "../gui/base/Button.js"
-import {copyToClipboard} from "../misc/ClipboardUtils.js"
-import {Icons} from "../gui/base/icons/Icons.js"
+import {Button, ButtonType} from "../../../gui/base/Button.js"
+import {copyToClipboard} from "../../../misc/ClipboardUtils.js"
+import {Icons} from "../../../gui/base/icons/Icons.js"
 import QRCode from "qrcode-svg"
-import {htmlSanitizer} from "../misc/HtmlSanitizer.js"
-import {Dialog, DialogType} from "../gui/base/Dialog.js"
-import {Icon, progressIcon} from "../gui/base/Icon.js"
-import {theme} from "../gui/theme.js"
-import type {U2fRegisteredDevice, User} from "../api/entities/sys/TypeRefs.js"
-import {createSecondFactor, GroupInfoTypeRef} from "../api/entities/sys/TypeRefs.js"
+import {htmlSanitizer} from "../../../misc/HtmlSanitizer.js"
+import {Dialog, DialogType} from "../../../gui/base/Dialog.js"
+import {Icon, progressIcon} from "../../../gui/base/Icon.js"
+import {theme} from "../../../gui/theme.js"
+import type {U2fRegisteredDevice, User} from "../../../api/entities/sys/TypeRefs.js"
+import {createSecondFactor, GroupInfoTypeRef} from "../../../api/entities/sys/TypeRefs.js"
 import {assertNotNull, LazyLoaded, neverNull} from "@tutao/tutanota-utils"
-import {locator} from "../api/main/MainLocator.js"
-import {getEtId, isSameId} from "../api/common/utils/EntityUtils.js"
-import {logins} from "../api/main/LoginController.js"
-import * as RecoverCodeDialog from "./RecoverCodeDialog.js"
-import {validateWebauthnDisplayName, WebauthnClient} from "../misc/2fa/webauthn/WebauthnClient.js"
-import {EntityClient} from "../api/common/EntityClient.js"
-import {ProgrammingError} from "../api/common/error/ProgrammingError.js"
+import {locator} from "../../../api/main/MainLocator.js"
+import {getEtId, isSameId} from "../../../api/common/utils/EntityUtils.js"
+import {logins} from "../../../api/main/LoginController.js"
+import * as RecoverCodeDialog from "../RecoverCodeDialog.js"
+import {validateWebauthnDisplayName, WebauthnClient} from "../../../misc/2fa/webauthn/WebauthnClient.js"
+import {EntityClient} from "../../../api/common/EntityClient.js"
+import {ProgrammingError} from "../../../api/common/error/ProgrammingError.js"
 import type {TotpSecret} from "@tutao/tutanota-crypto"
-import {IconButton, IconButtonAttrs} from "../gui/base/IconButton.js"
-import {ButtonSize} from "../gui/base/ButtonSize.js"
+import {IconButton, IconButtonAttrs} from "../../../gui/base/IconButton.js"
+import {ButtonSize} from "../../../gui/base/ButtonSize.js"
 
 const enum VerificationStatus {
 	Initial = "Initial",
