+++ src/settings/login/secondfactor/EditSecondFactorsForm.ts
@@ -1,25 +1,25 @@
 import m, {Children} from "mithril"
-import {assertMainOrNode, isTutanotaDomain} from "../api/common/Env"
-import type {User} from "../api/entities/sys/TypeRefs.js"
-import {SecondFactorTypeRef} from "../api/entities/sys/TypeRefs.js"
+import {assertMainOrNode, isTutanotaDomain} from "../../../api/common/Env.js"
+import type {User} from "../../../api/entities/sys/TypeRefs.js"
+import {SecondFactorTypeRef} from "../../../api/entities/sys/TypeRefs.js"
 import {assertNotNull, LazyLoaded, neverNull, ofClass} from "@tutao/tutanota-utils"
-import {Icons} from "../gui/base/icons/Icons"
-import {Dialog} from "../gui/base/Dialog"
-import {InfoLink, lang} from "../misc/LanguageViewModel"
-import {assertEnumValue, SecondFactorType} from "../api/common/TutanotaConstants"
-import {logins} from "../api/main/LoginController"
-import {appIdToLoginDomain} from "../misc/2fa/SecondFactorHandler"
-import {showProgressDialog} from "../gui/dialogs/ProgressDialog"
-import type {TableAttrs, TableLineAttrs} from "../gui/base/Table.js"
-import {ColumnWidth, Table} from "../gui/base/Table.js"
-import {NotFoundError} from "../api/common/error/RestError"
-import type {EntityUpdateData} from "../api/main/EventController"
-import {isUpdateForTypeRef} from "../api/main/EventController"
-import {ifAllowedTutanotaLinks} from "../gui/base/GuiUtils"
-import {locator} from "../api/main/MainLocator"
-import {EditSecondFactorDialog, SecondFactorTypeToNameTextId} from "./EditSecondFactorDialog"
-import {IconButtonAttrs} from "../gui/base/IconButton.js"
-import {ButtonSize} from "../gui/base/ButtonSize.js"
+import {Icons} from "../../../gui/base/icons/Icons.js"
+import {Dialog} from "../../../gui/base/Dialog.js"
+import {InfoLink, lang} from "../../../misc/LanguageViewModel.js"
+import {assertEnumValue, SecondFactorType} from "../../../api/common/TutanotaConstants.js"
+import {logins} from "../../../api/main/LoginController.js"
+import {appIdToLoginDomain} from "../../../misc/2fa/SecondFactorHandler.js"
+import {showProgressDialog} from "../../../gui/dialogs/ProgressDialog.js"
+import type {TableAttrs, TableLineAttrs} from "../../../gui/base/Table.js"
+import {ColumnWidth, Table} from "../../../gui/base/Table.js"
+import {NotFoundError} from "../../../api/common/error/RestError.js"
+import type {EntityUpdateData} from "../../../api/main/EventController.js"
+import {isUpdateForTypeRef} from "../../../api/main/EventController.js"
+import {ifAllowedTutanotaLinks} from "../../../gui/base/GuiUtils.js"
+import {locator} from "../../../api/main/MainLocator.js"
+import {EditSecondFactorDialog, SecondFactorTypeToNameTextId} from "./EditSecondFactorDialog.js"
+import {IconButtonAttrs} from "../../../gui/base/IconButton.js"
+import {ButtonSize} from "../../../gui/base/ButtonSize.js"
 
 assertMainOrNode()
